+++ varframe/__init__.py
@@ -92,4 +92,4 @@ __all__ = [
     "ModelList",
 ]
 
-__version__ = "1.2.1"
+__version__ = "1.3.0"